pyproject.toml
[project]
name = "jsonchema_ui"
version = "0.1.0"
description = "JSONSchema UI extraction"
readme = "README.rst"
authors = [
    { name = "Christian Klinger", email = "ck@novareto.de" }
]
requires-python = ">=3.11"
dependencies = ["pydantic >= 2", "colander", "deform"]

[project.optional-dependencies]
test = [
     "pytest",
     "PyHamcrest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["jsonschema_ui"]
namespaces = false